+++ bundle/config/services.php
@@ -9,13 +9,19 @@ return function (ContainerConfigurator $container): void {
         ->autoconfigure(true);
 
     // load all files as services
-    $services->load('Hyvor\\Internal\\Bundle\\', '../src');
+    $services
+        ->load('Hyvor\\Internal\\Bundle\\', '../src')
+        ->exclude([
+            '../src/Comms/Event',
+            '../src/Testing'
+        ]);
 
     $internalServices = $services
         ->load('Hyvor\\Internal\\', '../../src')
         ->exclude([
             '../../src/config.php',
-            '../../src/Auth/AuthUser.php',
+            '../../src/User/AuthUserOrganization.php',
+            '../../src/User/AuthUser.php',
             '../../src/Billing/Dto',
             '../../src/Billing/License',
             '../../src/Billing/Usage',
@@ -32,7 +38,7 @@ return function (ContainerConfigurator $container): void {
         $services->set(\Hyvor\Internal\Bundle\Mail\Component\Brand::class)->public();
     }
 
-//    $services->load('Hyvor\\Internal\\Auth\\', '../../src/Auth');
+//    $services->load('Hyvor\\Internal\\User\\', '../../src/User');
 //    $services->load('Hyvor\\Internal\\InternalApi\\', '../../src/InternalApi');
 //    $services->load('Hyvor\\Internal\\Util\\', '../../src/Util');
 };