+++ bundle/src/Comms/Event/FromCore/User/UserDeleted.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Hyvor\Internal\Bundle\Comms\Event\FromCore\User;
+
+use Hyvor\Internal\Bundle\Comms\Event\AbstractEvent;
+use Hyvor\Internal\Component\Component;
+
+class UserDeleted extends AbstractEvent
+{
+
+    public function __construct(
+        private int $userId
+    ) {}
+
+    public function getUserId(): int
+    {
+        return $this->userId;
+    }
+
+    public function from(): array
+    {
+        return [Component::CORE];
+    }
+
+    public function to(): array
+    {
+        return [];
+    }
+}