+++ bundle/src/Testing/ApiTestingTrait.php
@@ -7,8 +7,10 @@ use Symfony\Component\HttpFoundation\Response;
 trait ApiTestingTrait
 {
 
+    use BaseTestingTrait;
+
     /**
-     * @return array<string, mixed>
+     * @return array<mixed>
      */
     public function getJson(): array
     {
@@ -19,6 +21,17 @@ trait ApiTestingTrait
         $json = json_decode($content, true);
         $this->assertIsArray($json);
         return $json;
+    }
+
+    public function assertFailed(int $code, ?string $message = null): void
+    {
+        $response = self::getClient()->getResponse();
+        $this->assertSame($code, $response->getStatusCode());
+
+        if ($message !== null) {
+            $error = $this->getJson()['message'] ?? '';
+            $this->assertStringContainsString($message, $error);
+        }
     }
 
     public function assertViolationCount(int $count): void