+++ src/Billing/BillingFake.php
@@ -5,6 +5,7 @@ namespace Hyvor\Internal\Billing;
 use Hyvor\Internal\Billing\Dto\LicenseOf;
 use Hyvor\Internal\Billing\Dto\LicensesCollection;
 use Hyvor\Internal\Billing\License\License;
+use Hyvor\Internal\Billing\License\Resolved\ResolvedLicense;
 use Hyvor\Internal\Component\Component;
 use Hyvor\Internal\InternalConfig;
 use Symfony\Component\DependencyInjection\Container;
@@ -13,7 +14,7 @@ class BillingFake implements BillingInterface
 {
 
     /**
-     * @param License|(callable(int $userId, ?int $blogId, Component $component) : ?License)|null $license
+     * @param License|(callable(int $organizationId, ?int $blogId, Component $component) : ?License)|null $license
      * @param LicensesCollection|(callable(LicenseOf[] $of, Component $component) : LicensesCollection)|null $licenses
      * @return void
      */
@@ -31,7 +32,7 @@ class BillingFake implements BillingInterface
     }
 
     /**
-     * @param License|(callable(int $userId, ?int $blogId, Component $component) : ?License)|null $license
+     * @param License|(callable(int $organizationId, ?int $blogId, Component $component) : ?License)|null $license
      * @param LicensesCollection|(callable(LicenseOf[] $of, Component $component) : LicensesCollection)|null $licenses
      * @return void
      */
@@ -55,7 +56,7 @@ class BillingFake implements BillingInterface
         private InternalConfig $internalConfig,
 
         /**
-         * @param License|(callable(int $userId, ?int $resouceId, Component $component) : ?License)|null $license
+         * @param License|(callable(int $organizationId, ?int $resouceId, Component $component) : ?License)|null $license
          */
         private readonly mixed $license = null,
 
@@ -66,7 +67,7 @@ class BillingFake implements BillingInterface
     ) {
     }
 
-    public function license(int $userId, ?int $resourceId, ?Component $component = null): ?License
+    public function license(int $organizationId, ?Component $component = null): ResolvedLicense
     {
         $component ??= $this->internalConfig->getComponent();
 
@@ -78,13 +79,14 @@ class BillingFake implements BillingInterface
             return $this->license;
         }
 
-        return ($this->license)($userId, $resourceId, $component);
+        return ($this->license)($organizationId, $component);
     }
 
     /**
-     * @param array<LicenseOf> $of
+     * @param int[] $organizationIds
+     * @return array<int, ResolvedLicense>
      */
-    public function licenses(array $of, ?Component $component = null): LicensesCollection
+    public function licenses(array $organizationIds, ?Component $component = null): array
     {
         $component ??= $this->internalConfig->getComponent();