+++ src/Billing/License/Property/LicensePropertyNumberFormatting.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Hyvor\Internal\Billing\License\Property;
+
+enum LicensePropertyNumberFormatting: string {
+
+    case DEFAULT = 'default';
+    case BYTES = 'bytes';
+
+}