+++ src/Http/Middleware/AccessAuthUser.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Hyvor\Internal\Http\Middleware;
-
-use Hyvor\Internal\Auth\AuthUser;
-
-class AccessAuthUser extends AuthUser
-{}