+++ tests/Unit/Billing/LicenseTest.php
@@ -7,8 +7,11 @@ use Hyvor\Internal\Billing\License\CoreLicense;
 use Hyvor\Internal\Billing\License\DerivedFrom;
 use Hyvor\Internal\Billing\License\License;
 use Hyvor\Internal\Billing\License\PostLicense;
+use Hyvor\Internal\Billing\License\Property\LicensePropertyType;
 use Hyvor\Internal\Billing\License\TalkLicense;
+use Hyvor\Internal\Component\Component;
 use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 #[CoversClass(License::class)]
@@ -19,37 +22,82 @@ use PHPUnit\Framework\TestCase;
 class LicenseTest extends TestCase
 {
 
-    public function testCoreLicense(): void
+    /**
+     * @return class-string<License>[][]
+     */
+    public static function getAllLicenseClasses(): array
     {
-        $license = new CoreLicense();
-        $license->setDerivedFrom(DerivedFrom::CUSTOM_RESOURCE);
-        $this->assertEquals(DerivedFrom::CUSTOM_RESOURCE, $license->derivedFrom);
-        // add more tests when we have features
+        $classes = [];
+        foreach (Component::cases() as $component) {
+            $classes[] = [$component->license()];
+        }
+        return $classes;
     }
 
-    public function testBlogsLicense(): void
+    /**
+     * @param class-string<License> $licenseClass
+     */
+    #[DataProvider('getAllLicenseClasses')]
+    public function test_license_properties(string $licenseClass): void
     {
-        $license = new BlogsLicense();
+        $declaredProperties = $licenseClass::properties();
 
-        $this->assertEquals(2, $license->users);
-        $this->assertEquals(1_000_000_000, $license->storage);
+        $classReflection = new \ReflectionClass($licenseClass);
+        $propertyReflections = $classReflection->getProperties();
+
+        $this->assertTrue($classReflection->isFinal());
+
+        foreach ($propertyReflections as $propertyReflection) {
+
+            $key = $propertyReflection->getName();
+            $type = $propertyReflection->getType();
+
+            assert($type instanceof \ReflectionNamedType);
+            $typeName = $type->getName();
+
+            // check that the property is either int or bool
+            $this->assertTrue($typeName === 'int' || $typeName === 'bool');
+            $this->assertFalse($type->allowsNull());
+
+            $propertyType = LicensePropertyType::from($typeName);
+
+            // check that this property is declared
+            $declaredProperty = array_find($declaredProperties, fn($p) => $p->key === $key);
+            $this->assertNotNull($declaredProperty, 'all properties must be declared: ' . $key . ' missing in ' . $licenseClass);
+
+            $this->assertSame($propertyType, $declaredProperty->type);
+            $this->assertNotEmpty($declaredProperty->name);
+            $this->assertNotEmpty($declaredProperty->description);
+
+        }
+
+        $this->assertSame(
+            count($declaredProperties),
+            count($propertyReflections)
+        );
     }
 
-    public function test_talk_license(): void
+
+    public function test_build_from_array(): void
     {
-        $license = new TalkLicense();
-        $this->assertEquals(1_000, $license->credits);
-        $this->assertEquals(100_000_000, $license->storage);
-        $this->assertTrue($license->sso);
-        $this->assertFalse($license->noBranding);
-        $this->assertTrue($license->webhooks);
+
+        $license = BlogsLicense::fromArray(
+            // @phpstan-ignore-next-line
+            [
+                'users' => 1000,
+
+                // safely skips the following
+                'invalidkey' => 0, // invalid key
+                'aiTokens' => false, // invalid type
+                'analyses' => 30,
+                'badtype' => 'ohno'
+            ]
+        );
+
+        $this->assertInstanceOf(BlogsLicense::class, $license);
+        $this->assertSame(1000, $license->users);
+        $this->assertSame(false, $license->analyses);
     }
 
-    public function test_post_license(): void
-    {
-        $license = new PostLicense();
-        $this->assertEquals(0, $license->emails);
-        $this->assertFalse($license->allowRemoveBranding);
-    }
 
 }